Rename cocotb.sim_time_utils back to cocotb.utils
