+++ src/cocotb/utils.py
