+++ tests/pytest/test_timescale.py
@@ -15,14 +15,14 @@ from test_cocotb import (
     tests_dir,
 )
 
-from cocotb.sim_time_utils import _get_log_time_scale
+from cocotb.utils import _get_log_time_scale
 from cocotb_tools.runner import get_runner
 
 sys.path.insert(0, os.path.join(tests_dir, "pytest"))
 
 cocotb_test_contents = """
 import cocotb
-from cocotb.sim_time_utils import _get_simulator_precision
+from cocotb.utils import _get_simulator_precision
 
 @cocotb.test()
 async def check_timescale(dut):
