+++ tests/test_cases/test_async_bridge/test_async_bridge.py
@@ -31,8 +31,8 @@ import pytest
 import cocotb
 from cocotb.clock import Clock
 from cocotb.decorators import bridge
-from cocotb.sim_time_utils import get_sim_time
 from cocotb.triggers import ReadOnly, RisingEdge, Timer
+from cocotb.utils import get_sim_time
 
 
 def return_two(dut):
