+++ tests/test_cases/test_cocotb/test_clock.py
@@ -14,9 +14,9 @@ import pytest
 
 import cocotb
 from cocotb.clock import Clock
-from cocotb.sim_time_utils import get_sim_time
 from cocotb.simulator import clock_create, get_precision
 from cocotb.triggers import RisingEdge, Timer
+from cocotb.utils import get_sim_time
 
 LANGUAGE = os.environ["TOPLEVEL_LANG"].lower().strip()
