+++ tests/test_cases/test_cocotb/test_timing_triggers.py
@@ -20,7 +20,6 @@ import pytest
 
 import cocotb
 from cocotb.clock import Clock
-from cocotb.sim_time_utils import get_sim_time
 from cocotb.simulator import get_precision
 from cocotb.triggers import (
     First,
@@ -32,6 +31,7 @@ from cocotb.triggers import (
     Timer,
     with_timeout,
 )
+from cocotb.utils import get_sim_time
 
 LANGUAGE = os.environ["TOPLEVEL_LANG"].lower().strip()
 SIM_NAME = cocotb.SIM_NAME.lower()
